feat(deposit-validator): separate from panel main component
